+++ api-sdk/package.json
@@ -15,21 +15,12 @@
   "dependencies": {
     "axios": "^0.20.0",
     "highlight.js": "^10.3.1",
-    "markdown-it": "^12.0.1",
-    "markdown-it-abbr": "^1.0.4",
-    "markdown-it-anchor": "^6.0.0",
-    "markdown-it-deflist": "^2.1.0",
-    "markdown-it-footnote": "^3.0.2",
-    "markdown-it-ins": "^3.0.0",
-    "markdown-it-mark": "^3.0.0",
-    "markdown-it-sup": "^1.0.0",
-    "markdown-toc": "^1.2.0"
+    "markdown-toc": "^1.2.0",
+    "remarkable": "^2.0.1"
   },
   "devDependencies": {
     "@types/chai": "^4.2.13",
     "@types/highlightjs": "^9.12.0",
-    "@types/markdown-it": "^10.0.2",
-    "@types/markdown-it-anchor": "^4.0.4",
     "@types/mocha": "^8.0.3",
     "@types/node": "^14.14.2",
     "chai": "^4.2.0",
